refactor: light attenuation 변경 및 type safety 보완
- Changed default `AttenuationRadius` in `LocalLightComponent.h` from `30.0f` to `5.0f` to reduce light range.
- Updated `SetMaterial` function in `StaticMeshComponent.cpp` to use `static_cast<uint32>` for better type compatibility.

+++ Mundi/Source/Runtime/Engine/Components/StaticMeshComponent.cpp
@@ -348,7 +348,7 @@ void UStaticMeshComponent::SetMaterialByUser(const uint32 InMaterialSlotIndex, c
 
 void UStaticMeshComponent::SetMaterial(uint32 InElementIndex, UMaterial* InNewMaterial)
 {
-	if (0 <= InElementIndex && InElementIndex < MaterialSlots.Num())
+	if (0 <= InElementIndex && InElementIndex < static_cast<uint32>(MaterialSlots.Num()))
 	{
 		MaterialSlots[InElementIndex] = InNewMaterial;
 	}
